requirements.txt
# Core LangChain
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2

# Vector Store
faiss-cpu==1.7.4

# Document Loaders
pytube==15.0.0
youtube-transcript-api==0.6.1
PyPDF2==3.0.1
pypdf==3.17.4
wikipedia==1.4.0

# OpenAI
openai==1.6.1

# Text Processing
tiktoken==0.5.2
unstructured==0.11.6

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: AWS Transcribe (if using)
boto3==1.34.10
